API Documentation

Below is the API documentation for Mixtape.jl

Interception interfaces

These interfaces parametrize the Mixtape pipeline and allow you to transform lowered code, optimize Core.Compiler.IRCode, and turn on debug printing. Similar to Cassette.jl – to override these interfaces, users subtype CompilationContext.

Mixtape.CompilationContext — Type

abstract type CompilationContext end

Parametrize the Mixtape pipeline by inheriting from CompilationContext. Similar to the context objects in Cassette.jl. By using the interface methods show_after_inference, show_after_optimization, debug, transform, preopt!, and postopt! – the user can control different parts of the compilation pipeline.

Mixtape.allow — Function

allow(f::CompilationContext, args...)::Bool

Determines whether the user-defined transform, preopt!, and postopt! are allowed to look at a lowered CodeInfoTools.Pipe object or Core.Compiler.IRCode.

The user is allowed to greenlight modules:

allow(::MyCtx, m::Module) == m == SomeModule

or even specific signatures

allow(::MyCtx, fn::typeof(rand), args...) = true

Mixtape.transform — Function

transform(ctx::CompilationContext, b::Core.CodeInfo)::Core.CodeInfo
transform(ctx::CompilationContext, b::Core.CodeInfo, sig::Tuple)::Core.CodeInfo

User-defined transform which operates on lowered CodeInfo in the form of a CodeInfoTools.Pipe object. There's two versions: (1) ignores the signature of the current method body under consideration and (2) provides the signature as sig.

Transforms might typically follow a simple "swap" format using CodeInfoTools.Builder:

function transform(::MyCtx, src)
    b = CodeInfoTools.Builder(b)
    for (k, st) in b
        b[k] = swap(st))
    end
    return CodeInfoTools.finish(b)
end

but more advanced formats are possible. For further utilities, please see CodeInfoTools.jl.

Mixtape.preopt! — Function

preopt!(ctx::CompilationContext, ir::Core.Compiler.IRCode)::Core.Compiler.IRCode

User-defined transform which operates on inferred Core.Compiler.IRCode. This transform operates before a set of optimizations which mimic Julia's pipeline.

Mixtape.postopt! — Function

postopt!(ctx::CompilationContext, ir::Core.Compiler.IRCode)::Core.Compiler.IRCode

User-defined transform which operates on inferred Core.Compiler.IRCode. This transform operates after a set of optimizations which mimic Julia's pipeline.

Mixtape.show_after_inference — Function

show_after_inference(ctx::CompilationContext)::Bool

Turns on a pipeline feature which will dump out CodeInfo after inference (including the user-defined transform transformation, if applied).

Mixtape.show_after_optimization — Function

show_after_optimization(ctx::CompilationContext)::Bool

Turns on a pipeline feature which will dump out CodeInfo after optimization (including the user-defined preopt! and postopt! transformations, if applied).

Mixtape.debug — Function

debug(ctx::CompilationContext)::Bool

Turn on debug tracing for inference and optimization. Displays an instrumentation trace as inference and optimization proceeds.

Mixtape.@ctx — Macro

@ctx(properties, expr)

Utility macro which expands to implement a subtype of CompilationContext. Also allow the user to easily configure the debug static tracing features and show_after_inference/show_after_optimization Boolean function flags.

Usage:

@ctx (b1::Bool, b2::Bool, b3::Bool) struct MyCtx
    fields...
end

Expands to:

struct MyCtx <: CompilationContext
    fields...
end
show_after_inference(::MyCtx) = b1
show_after_optimization(::MyCtx) = b2
debug(::MyCtx) = b3

Call interfaces

These interfaces allow you to apply the Mixtape pipeline and execute code with a linked in OrcJIT instance through GPUCompiler.jl.

Mixtape.jit — Function

jit(ctx::CompilationContext, f::F, tt::TT = Tuple{}) where {F, TT <: Type}

Compile and specialize a method instance for signature Tuple{f, tt.parameters...} with pipeline parametrized by ctx.

Returns a callable "thunk" Entry{F, RT, TT} where RT is the return type of the instance after inference.

Mixtape.@load_call_interface — Macro

@load_call_interface()

A macro which expands to load a generated function call into the scope of the calling module. This generated function can be applied to signature argument types Tuple{ctx<:CompilationContext, f<:Function, args...}. call then creates a new instance of ctx and calls Mixtape.jit – it then caches a ccall which calls a function pointer to the GPUCompiler-compiled LLVM module.

The call interface currently uses a slow ABI ccall – which costs an array allocation for each toplevel call. This allocation is required to construct a Vector{Any} for the arguments and pass a pointer to it over the line, where the call unboxes each argument.